requirements-dev.txt
# Development dependencies for Control4 Dimmers
homeassistant
paho-mqtt
